float used instead of int for radius.

+++ Twist/src/com/secondhand/model/Enemy.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.math.Vector2;
 public class Enemy extends BlackHole {
 	private boolean isBigger;
 	
-	public Enemy (Vector2 vector, int radius) {
+	public Enemy (Vector2 vector, float radius) {
 		super(vector,radius);
 	}
